Clarify why `cfg` is mandatory for `executable=True` attributes.
I was originally considering removing `cfg = "target"` support in the name
of API cleanup, since that's a semantic no-op. But the reasoning is fair
for when `cfg` is mandatory (shared in a Google doc by laurentlb@):

"When an input is executable, we often want to run it on the host. So I
suggest that cfg is mandatory when input is executable".

But I don't think we should set `cfg = "target"` casually when it's not
helpful - that'll just litter code bases. I tried to address these issues
in this PR.

Also note, for posterity, that `cfg = "host"` is due for deprecation:
https://github.com/bazelbuild/proposals/blob/master/designs/2019-02-12-execution-transitions.md