Avoid passing Tristate and Distrib type attributes to attr struct
These types are not Starlark-compatible, and, due to an existing bug, these are propagated anyway for native rule targets which include them as implicit attributes. Without this change, one may assign to these types in Starlark but later get a catastrophic error when an object of these types is passed to another function, such as dir().

Progress toward #7281.

RELNOTES: None.
PiperOrigin-RevId: 243336671