Report the full chain of repositories for WORKSPACE cycles
When a cycle through the workspace file is found, this should indicate that
a repository was used prior to being defined (either because it was defined
too late, or because it wasn't defined at all). In this case do not only
report the load statement that eventually lead to the missing definition,
but the full list of repositories involved; in particular, indicate that
the last one in this chain is the one where the definition is missing.

Fixes #7871.
Improves on #7784.

Change-Id: Ic3eb499460c87ad5c78f21373d34e2ccbba2f2fe
PiperOrigin-RevId: 243804742