Make sandbox symlink forest deletions asynchronous.
Each sandbox action runs within a symlink forest that exists in a separate
subtree because we use a unique identifier for those subtrees.  Therefore
it is unnecessary to delete those trees in the critical path.

Tree deletions can be very expensive, especially on macOS, so make them
asynchronous if --experimental_sandbox_async_tree_delete_idle_threads is
given.  When this flag is not zero, Bazel will schedule all deletions on
a separate low-priority thread while the build is running, and will then
use the requested number of threads once the build is done to quickly
catch up with an still-ongoing deletions.

For a large iOS build, this cuts down clean build times with sandboxed
enabled significantly.  Helps more on machines with more cores:

* On a Mac Pro 2013, the improvement is almost 20%:

  standalone:      mean 2746.33, median 2736.00, stddev 33.07
  sandboxed-async: mean 4394.67, median 4393.00, stddev 33.09
  sandboxed-sync:  mean 5284.33, median 5288.00, stddev 20.17

* On a MacBook Pro 2015, we see a more modest 10% improvement:

  standalone:      mean 3418.33, median 3422.00, stddev 7.41
  sandboxed-async: mean 5090.00, median 5086.00, stddev 40.92
  sandboxed-sync:  mean 5694.67, median 5700.00, stddev 37.75

Partially addresses https://github.com/bazelbuild/bazel/issues/7527.

RELNOTES: None.
PiperOrigin-RevId: 243805556