Add build option --repo_env
Add an option to environment variables only for repository rules. This
change is not about making more environment variables available to repository
rules, as they have access to the whole environment anyway. However, having
an option to set environment variables allows to use the config mechanism
of .bazelrc to pass information to external repositories without invalidating
the whole action graph.

This way of using external repositories to change sources according to the used
configuration is might not be in line with the idea of bazel to be able to build
for all configurations from the same source tree, but there is sufficient demand
for such a feature.

Change-Id: Ib312958f5faa055863f1cd643efbcff63d0b16a3
PiperOrigin-RevId: 243810658