Invalidate all paths from a failed action to its lost input in rewinding
Rewinding must invalidate all Skyframe paths from the failed action to
the action which generates the lost input. Intermediate nodes not on the
shortest path to that action may have values that depend on the output
of that action. If these intermediate nodes are not invalidated, then
their values may become stale.

RELNOTES: None.
PiperOrigin-RevId: 243835353