Finish node completion when deps re-complete during registration
Just before completion, a parent node registers itself as an rdep of the
deps newly discovered during its final restart.

If one of these deps was rewound between when its value was consumed and
when the parent registers as an rdep, then that dep may be scheduled for
evaluation, and will signal the parent when it completes.

It's possible for such a dep to send that signal before the parent
finishes registering itself with other deps. If this happens, then the
parent could ready itself for completion with the last signalDep call
during its dep registration. And if this happens, the parent should
finish its completion, because nothing else will schedule it.

RELNOTES: None.
PiperOrigin-RevId: 243904314