C++: Changes internal C++ compile() API to match Bazel
This is the new API introduced in https://github.com/bazelbuild/bazel/commit/0b03a28701d297c9ffe976c54580ceff32eec1b7. In this CL we keep the
SkylarkCallables separate so that the internal rules work with the old and the
new API to facilitate migration. link() will be done in a separate CL.

RELNOTES:none
PiperOrigin-RevId: 243941778