Unpack and repack Bazel's deploy jar without compression.
This reduces final binary size by ~10%.
Binary size on linux is now 61MB.

This comes with the downside that the install size of Bazel grows by
~25MB to 184MB. But given the fact that it was 221MB with Bazel 0.23
this seems to be acceptible.
We also have further plans to reduce the binary size even more which
will also help the install base size (e.g. proguarding).

RELNOTES: None