Skylark repositories: also record the definition of the rule
When fetching a repository fails, in particular due to an unmet dependency,
it is often also helpful to know the place where the repository rule was defined,
e.g., to find some hints on why an implicit dependency was added. So record that
information as well in the definition information shown upon failures related to
that particular repository.

Fixes #7784.

Change-Id: Ie2be1b1e48283b4b4199bfb216011d935c5250a1
PiperOrigin-RevId: 243962184