Upon noticing an error in an external repository, also report the reason
Users sometimes find it hard (issue #7850) to identify the root
cause of a failure, if it is a fetch in an external repository.
As we print a message anyway on observing a failure in a fetch,
make this message more helpful by
- upgrading it to a proper error message, and
- including the message of the actual fetch error.
Also ensure that in case of a repository fetch failing due to
the execution of `fail` in the implementation function, the
message is reported as well.

Fixes #7850
Fixes #3683

Change-Id: I70eea4f966c98949e0f57462ac73b2a078552073
PiperOrigin-RevId: 243988092