Delete BuildEventServiceAbruptExitCallback and handle BEP upload exceptions in the BES module.
Since the closing of all the BEP transports now happens in the BES module we don't need to use the BuildEventServiceAbruptExitCallback, instead we make use of SettableFuture.setException to signal a problem with the upload. This makes the code (a lot) cleaner by removing one layer of indirection for the error handling of the BEP uploads.

As a consequence of this change, all BES failures are now reported in afterCommand (or beforeCommand for the async case) which means that, for instance, if the connection with BES times out we'll wait until the build is finished to report the time out, instead of doing it mid-build. Exiting early was a leftover of the original design of the BES code but given the amount of complexity it adds and the little benefit we've decided to drop it.

PiperOrigin-RevId: 244064861