Windows, JNI: WaitableProcess takes relative argv0
WaitableProcess::Create's argv0 argument may now
be relative.

If argv0 is relative, and cwd has a value, then we
make argv0 relative to cwd.

If argv0 is relative, and cwd is empty, we make
argv0 relative to the current directory of the
process.

This change allows using WaitableProcess in
blaze_util_windows.cc in ExecuteProgram, which is
sometimes called with a relative path.