Windows, file.cc: Make DeletePath simpler and more robust
Old behavior:

First try to delete the path by DeleteFileW, if it fails, check the error code and the file attribute to tell what's the reason. Then if it's a directory, try to delete it with RemoveDirectoryW, and if it's a file, try to remove read-only attribute and delete file again.

This is sometimes flaky because the GetFileAttributesW function could fail with access denied error after the first DeleteFileW call if the file is only marked as deleted but not yet deleted.

New behavior:

1. Check the file attribute first to see if we should use DeleteFileW (for file and symlink) or RemoveDirectoryW (for directory and junction).
2. If it's a directory, try to delete it in a robust way (check the directory status first, retry if necessary)
3. If it's a file, remove the read-only attribute if necessary and try to delete it.

RELNOTES: None
PiperOrigin-RevId: 244188343