Add short circuiting returns to equals method in OptionsBase.
asMap() is expensive as it reflectively aggregates all fields into a map. We can return quickly if they are the same instance or instances of differing classes. If we do get to comparing different instances of the same class, we can return as soon as we find a differing option field.

RELNOTES: None.
PiperOrigin-RevId: 244687761