Reset platform on cpu transitions.
With platform mappings about to be enabled we need to make sure that platforms are reset in any configuration transitions that write flags which may impact the platform. Otherwise a platform may be set before the transition that matches the old flags but platform mapping would not be triggered for the new flag values (because platform mapping only happens if no platform is set).

Step 6/N towards the platforms mapping functionality for https://github.com/bazelbuild/bazel/issues/6426

RELNOTES: None.
PiperOrigin-RevId: 244696779