Fix flakiness of ByteStreamUploaderTest#failedUploadsShouldNotDeduplicate
The ByteStreamUploader has logic to internally deduplicate uploads. The test first triggers an upload that fails followed by an identical upload that succeeds. It asserts that the failed upload is not cached so that the subsequent upload is actually performed.

For a given file the logic for state transition of "being uploaded" to "uploaded/failed" is done in a listener of the upload future. This listener is executed *after* the future completes (i.e. after get()/isDone() returns true) and thus there would be a race between the second upload starting and the above mentioned state transition being executed. If the race was lost it would look to the second upload as if the first hadn't completed yet.

The fix is to introduce a future that only completes after the internal state tracking has been done.

RELNOTES: None
PiperOrigin-RevId: 245006516