Update Truth to 0.44.
This is necessary for some forthcoming changes that migrate to methods introduced in that version:
https://github.com/google/truth/releases/tag/release_0_44

While here, move diffutils from runtime_deps to deps. This is unlikely to matter in practice, but in principle it's the right thing, and it *might* matter if static analysis or other build tools (-Xlint:all?) went digging into the deps of Truth at compile time. Such problems (albeit when the dependency's API is publicly visible through the using library) are what prompted the creation of java_import#deps (internal Google bug 23781162). For other problems (that might also not apply here), see https://github.com/google/guava/issues/2721