Embed the java_tools BUILD file in the java tools archive.
Also add a corresponding `java_toolchain` target in the BUILD file.

The advantage of this change is that every java tools release will come with its own embedded toolchain. That allows us to remove all the targets under `@bazel_tools//tools/jdk`. Additionally it allows straightforward definitions of `toolchain`s, so we can platformize the Java rules in Bazel.

Closes #8130.

PiperOrigin-RevId: 245232441