De-flake //src/test/java/com/google/devtools/build/lib/buildeventstream/transports:BuildEventTransportTest
BinaryFormatFileTransportTest was making a series of assertions about the internal state of the FileTransport (more specifically the SequentialWriter) which was causing some flakiness as it asserted on a temporary state of the transport. Most of those assertions aren't really useful since they're not testing the public API of the class but instead it's testing implementation details.

With this change we remove the offending assertions and add the visibility to the members/methods of FileTransport. Because of this JsonFormatFileTransportTest needs to be updated too (which arguably should be the way it should have been done in the first place).

$ bazel test //src/test/java/com/google/devtools/build/lib/buildeventstream/transports:BuildEventTransportTest --runs_per_test=1000
//src/test/java/com/google/devtools/build/lib/buildeventstream/transports:BuildEventTransportTest PASSED in 15.1s

PiperOrigin-RevId: 245268064