Command Execution: distinguish aborted processes
In some cases, bazel terminates a process, e.g., because an error
was found in some other part of the build. If this happens for
a process called from a Skylark repository rule, we want to abort
the rule here, and not hand control back the to the caller; on
the other hand, we want to give the caller the chance to handle
all other kinds of abnormal termination, as executed commands
might be used to probe for properties of the underlying system.

Unfortunately, we cannot do this distinction perfectly, as for a
process terminated by signal 15, we do not know who sent that signal.
However, handling those cases as aborted should not interfere with
existing uses, and there's always a workaround (like a wrapper
process).

Fixes #7993.

Change-Id: Ibd9630b38103094e56570d9bba2f67a4197c18c5
PiperOrigin-RevId: 245381697