HttpConnector: do not report connection timeouts as user interrupts
If a conenct() times out (because the server is not responsive), it
throws a SocketTimeoutException which is an InterruptedIOException.
The latter, however, are mapped by the HttpConnector to InterruptedExceptions
indicating a user interruption of the download, which is generally correct, but
not for this specific subclass. Fix this by handling this subclass separately
and rethrowing it as ordinary IOException.

Fixes #3851.
Closes #6209.

Change-Id: I0c3b46921a725f5a5a728a83dd75c7c83c6a46f4
PiperOrigin-RevId: 245398269