http.bzl: Use syntactically correct checksums in examples
The Starlark version of http_archive comes with some documenting examples,
that also provide the sha256 parameter, as is best practice. However, the
example value is not syntactically a sha256 sum (in fact, it has to the
length of a sha1 sum) which can be irritating. Therefore, replace it with
a proper sha256 sum (that of the empty string).

Change-Id: I6ea7b4aa0f8b3603b670400650ca64cd8082cf61
PiperOrigin-RevId: 245410913