Extract runtime desugar lib dependencies to @android_tools
Context for kmb@: We're extracting runtime dependencies of the Android rules out into a separately hosted @android_tools tarball to keep the core Bazel binary small. This is a no-op change to desugaring, but it also means that future updates to desugar libs will require a new @android_tools release, and Bazel itself updated to use this new release.

This is an interim state where we have native Android rules and increasingly more tools living externally. The end goal is to extract all Android rules and tools and fetch them at build time only when building Android targets.

Before:

```
$ ls -al bazel-bin/src/bazel
-r-xr-xr-x 1 jingwen primarygroup 52307083 Apr 25 15:33 bazel-bin/src/bazel
```

After:

```
$ ls -al bazel-bin/src/bazel
-r-xr-xr-x 1 jingwen primarygroup 51351230 Apr 25 15:33 bazel-bin/src/bazel
```

The new filesize after this change is 49MiB.

Closes #8149.

Change-Id: Ic728d5e6760a8dda535d478e419626a1a40a3324
PiperOrigin-RevId: 245441732