Download stderr/stdout to a temporary FileOutErr
On Windows, we cannot delete file without closing the FileOutputStream, but we may need to write to the FileOutErr in fallback execution, so we cannot close the OutputStreams.

By using a temporary child FileOutErr, we can safely close and delete the stdout/stderr files of it when the download fails and the original FileOutErr will still be writable.

Fixes https://github.com/bazelbuild/bazel/issues/8104

RELNOTES: None
PiperOrigin-RevId: 245731397