Add the glob option `allow_empty`.
When it's set to False, the glob will fail if any included pattern matches nothing, or if the whole glob returns nothing.

The intent is to eventually have `allow_empty` to False by default to catch user errors.

RELNOTES: The glob function has a new argument `allow_empty`. When set to False, the glob fails when it doesn't match anything.
PiperOrigin-RevId: 245762652