SkylarkExecutionResult: gracefully handle missing stdout/stderr
If for a command stdout and/or stderr are not available, CommandResult
returns a special, non-public, object NO_OUTPUT_COLLECTED instead. When
passing that information up to Starlark in order to allow a rule to handle
it, map that to the empty string as the best description of the missing
output we can give.

Change-Id: I7c895b765a03bbc222edd254c81077eec94721f0
PiperOrigin-RevId: 245775743