Windows: open files with at least read-sharing
If a process opens files without read-sharing
(FILE_SHARE_READ) and that file is symlink or
junction, then other processes can't check whether
the target of it exists.

In particular, if a process opens a file with
CreateProcessW(dwShareMode = 0) then
symlink-following file existence checks (such as
java.nio.file.Files.exists and
java.nio.file.Files.notExists) fail to check if
the symlink's target exists (even if the file is
not a symlink).

Fixes https://github.com/bazelbuild/bazel/issues/7392