Always report when a package is done to the PackageProgressReceiver.
Previously, if a package was so broken that a even a errored PackageValue couldn't be created for it, it would hang around in the UI ("currently loading //some/broken/package") for the entire analysis-loading phase. This behavior was because PackageProgressReceiver didn't receive a done notification from PackageFunction in extreme error cases. The fix is to call PackageProgressReceiver.doneReadPackage even when exceptions happen during loading.

While we're here, save a level of indentation in PackageFunction.loadPackage() by moving the cache checking logic to the single caller.