Add a helper for comparing Versions as strictly lower
Version has a partial order defined by atMost method, which returns true if the
versions are comparable and left is lower or equal than the right one. Add a
convenience method, derived from atMost which check if 2 versions are
comparable and one is strictly lower than the other

RELNOTES: None.
BUG=130288327
PiperOrigin-RevId: 247098205