Fail the build if runs_per_test is greater than 100000.
The flag --runs_per_test directly affects the size of the BEP's target_complete event. The idea is that there will be one child event (test_result) for every run, therefore setting a very large --runs_per_test creates a target_complete event that can be too large for the BES backends to process.

Example:
$ bazel build //myproject:label --runs_per_test=100000 --build_event_json_file=/tmp/bep.json
$ cat /tmp/bep.json | grep '"id":{"targetCompleted"'
{"id":{"targetCompleted":{"label":"//myproject:label","configuration":{"id":"8d35b05fd34b37dd2b7bd9565d248570"}}},"children":[{"testResult":{"label":"//myproject:label","run":1,"shard":1,"attempt":1,"configuration":{"id":"8d35b05fd34b37dd2b7bd9565d248570"}}},{"testResult":{"label":"//myproject:label","run":2,"shard":1,"attempt":1,"configuration":{"id":"8d35b05fd34b37dd2b7bd9565d248570"}}},{"testResult":{"label":"//myproject:label","run":3,"shard":1,"attempt":1,"configuration":{"id":"8d35b05fd34b37dd2b7bd9565d248570"}}},{"testResult":{"label":"//myproject:label","run":4,"shard":1,"attempt":1,"configuration":{"id":"8d35b05fd34b37dd2b7bd9565d248570"}}},{"testResult":{"label":"//myproject:label","run":5,"shard":1,"attempt":1,"configuration":{"id":"8d35b05fd34b37dd2b7bd9565d248570"}}},{"testResult":{"label":"//myproject:label","run":6,"shard":1,"attempt":1,"configuration":{"id":"8d35b05fd34b37dd2b7bd9565d248570"}}},{"testResult":{"label":"//myproject:label","run":7,"shard":1,"attempt":1,"configuration":{"id":"8d35b05fd34b37dd2b7bd9565d248570"}}},
(... up to 100000 runs ...)

Notice that this change isn't incompatible since a high runs_per_test currently fails to upload events to BES given that the events are too large, so we're simply adding a better way of handling this situation without failing the build.

PiperOrigin-RevId: 247234673