Remove artificial parens in bazel query's BinaryOperationExpression#toString().
The previous #toString() implementation renders any number of operands as the
left-recursive parse tree creating the expression. For example:
`(a + b + c + d)`.toString() produced `(((a + b) + c) + d)`.

This presents an obstacle to logging large queries in human-facing log files,
because a query summing 100 labels will start with 100 open-parens and use 200
bytes for parens. At the long tail this can cause logged queries to be
truncated that would otherwise be preserved.

Note that the previous toString() result is still available by creating nested
BinaryOperationExpressions corresponding to the desired parenthesis grouping.

RELNOTES=none
PiperOrigin-RevId: 247332869