resolved file: include tool chains
When the repository resolved file was originally designed, the
WORKSPACE file would only declare repositories. Later, an additional
command `register_toolchains` was added. This broke the property
that a resolved file could be used a full replacement for a
WORKSPACE file. Restore this property by adding information about
the registered tool chains to the resolved file as well.

Change-Id: Ic23bfb91af961a0c7a4273c3998ad3adf2b4599c
PiperOrigin-RevId: 247383213