Run tests after building java tools.
This PR adds a shell integration test for the java rules that is run with the java toolchain of the java tools build from head. The test is created and run by `src/upload_java_tools.sh` which is used by the Buildkite pipeline to generate the java tools zip.

The generated java tools zip is uploaded to GCS only if the test succeeded. This provides testing before creating release candidates for the java tools.

Closes #8205.

PiperOrigin-RevId: 247408063