Canonicalise install path for comparison
Currently, if the install directory is placed on a path with a symlink
in it, this function will always believe there is a new version,
because it compares realpath() of the symlink with its intended
resolution.

This change realpath()s both the symlink and its intended resolution,
which should give correct behaviour even in the presence of symlinks.