Improve the documentation on unquoted words in query expressions
This mentions that $ and @ are allowed in unquoted words (previously omitted). The documentation of the characters allowed in unquoted words is also changed to list punctuation characters in the same relative order and format as the list of punctuation characters allowed in target names in that section of the documentation.

It also lists the full set of characters permitted in target expressions which are not permitted in unquoted words (the documentation previously mentioned + and = only in examples and ~ not at all). It emphasizes the implication of this: When writing code that generates query expressions, target names should be quoted.

A slight tweak is made to the bit about alphanumeric characters, since A-Za-z is not "the set of alphabetic characters" and 0-9 is not "numerals" unless you specify "in ASCII".

RELNOTES: None.
PiperOrigin-RevId: 247435415