Fix decoding of jar file paths for coverage runs of java
File paths that contain special characters (such as % or @) will be escaped using the URL encoding.

The previous implementation did not decode these paths before passing it to the file system, leading to
file-not-found errors.

This bug was triggered by having a maven dependency being placed in the following path (note the %40):

```
$HOME/.cache/bazel/_bazel_patrick/$HASH/external/maven/v1/https/user%40repo.provider.com/
```

Closes #8270.

PiperOrigin-RevId: 247572901