Windows, bazel run: don't join all args into argv0
To "bazel run" a binary, the RunCommand creates an
ExecRequest and sends it to the Bazel client,
which then runs the binary as the client's
subprocess.

In PR #8241, RunCommand merged all command line
arguments and added that to the request as a
single argv entry. The client uses the first argv
element as the executable, but if all arguments
are joined into one string then argv0 contains the
whole argument vector as a single string.

Unfortunately the test didn't catch this because
it didn't attempt passing any arguments.

Fixes https://github.com/bazelbuild/bazel/issues/8229
See https://github.com/bazelbuild/bazel/issues/8240

Closes #8287.

PiperOrigin-RevId: 247590904