Homogenize LocalEnvProvider instantiation in a single place.
This adds a new forCurrentOS factory method to LocalEnvProvider to construct
the correct local env provider for the current OS, and removes duplicate (and
inconsistent) logic from a bunch of places.

To make this change simpler, this also moves XcodeLocalEnvProvider from an
"apple" package (which is useless because it only contains one class and we
depend on in anyway in a Bazel build) to the "local" subpackage.

And as a side-effect, this gets rid of XcodeLocalEnvProviderTest, which I
don't understand how it could possibly have ever worked: the test was trying
to validate a condition on non-Darwin platforms but was later restricting
itself to running on Darwin platforms only.  We didn't see this because
this test was not being executed in the Bazel exported tree (that is, was
not run on BazelCI) and is now run after the code move.  I tried to fix the
test, but it's so out of sync with the current code that even understanding
what condition it wanted to test was hard (but it was not that useful).

RELNOTES: None.
PiperOrigin-RevId: 247663150