Query performance: resolve simple patterns directly
Instead of depending on the TargetPatternFunction for simple patterns,
depend directly on the PackageValue and use that to resolve the targets
outside of Skyframe.

The problem with TPF is that it returns labels rather than targets, so
query has to turn the labels back to targets in order to be able to
process the results.

For a synthetic benchmark, we're seeing a significant improvement in
performance.

Before:
Cold 0m18.469s
Hot   0m6.045s

After:
Cold 0m15.778s
Hot   0m4.892s
PiperOrigin-RevId: 248311593