Update getHost methods to also copy host flags, for use in exec->host transitions.
Currently, the exec transition also calls getHost. As currently implemented, this means any host_* options get reset to defaults, instead of what was actually passed. In a further transition to a real host config, the expected set of flags won't be set.

This change fixes getHost so that host_* options are kept, so that exec->host transitions work as expected.

Part of work on execution transitions, #7935.

PiperOrigin-RevId: 248386317