Don't process both ObjcProvider and CcInfo from the same dependency
ObjcLibrary exports both an ObjcProvider and a CcInfo.  The CcInfo is
completely redundant, but a dependent ObjcLibrary was processing both.
This interferes with the direct fields that we recently added.  This
change will cause the dependent provider to process only
oneof(ObjcProvider, CcInfo) per dependence, giving precedence to
ObjcProvider.

RELNOTES:None
PiperOrigin-RevId: 248631112