Add support for a cc_import() of a DLL with no interface library on Windows.
Depending on this is very nearly a no-op; the linker cannot reference DLLs with no interface library. However, it's necessary for the copy_dynamic_libraries_to_binary feature to work in cases where a DLL is referenced without use of its interface library. (e.g., LoadLibrary, or one DLL a target depends on that itself depends on another DLL).

RELNOTES: Allow cc_import() of a DLL with no interface library on Windows, used to document runtime dependencies.
PiperOrigin-RevId: 248716186